Move helper classes to their own headers
