+++ src/tests/profiler/native/getappdomainstaticaddress/getappdomainstaticaddress.h
@@ -14,68 +14,11 @@
 #include <string>
 #include <thread>
 #include <chrono>
-#include <condition_variable>
 #include <functional>
 #include "cor.h"
 #include "corprof.h"
 
 typedef HRESULT (*GetDispenserFunc) (const CLSID &pClsid, const IID &pIid, void **ppv);
-
-class AutoEvent
-{
-private:
-    std::mutex m_mtx;
-    std::condition_variable m_cv;
-    bool m_set = false;
-
-    static VOID DoNothing()
-    {
-
-    }
-
-public:
-    AutoEvent() = default;
-    ~AutoEvent() = default;
-    AutoEvent(AutoEvent& other) = delete;
-    AutoEvent(AutoEvent &&other) = delete;
-    AutoEvent &operator=(AutoEvent &other) = delete;
-    AutoEvent &operator=(AutoEvent &&other) = delete;
-
-    void Wait(std::function<void()> spuriousCallback = DoNothing)
-    {
-        std::unique_lock<std::mutex> lock(m_mtx);
-        while (!m_set)
-        {
-            m_cv.wait(lock, [&]() { return m_set; });
-            if (!m_set)
-            {
-                spuriousCallback();
-            }
-        }
-        m_set = false;
-    }
-
-    void WaitFor(int milliseconds, std::function<void()> spuriousCallback = DoNothing)
-    {
-        std::unique_lock<std::mutex> lock(m_mtx);
-        while (!m_set)
-        {
-            m_cv.wait_for(lock, std::chrono::milliseconds(milliseconds), [&]() { return m_set; });
-            if (!m_set)
-            {
-                spuriousCallback();
-            }
-        }
-        m_set = false;
-    }
-
-    void Signal()
-    {
-        std::unique_lock<std::mutex> lock(m_mtx);
-        m_set = true;
-        m_cv.notify_one();
-    }
-};
 
 class GetAppDomainStaticAddress : public Profiler
 {
